% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abagenes.R
\docType{data}
\name{abagenes}
\alias{abagenes}
\title{Data from the Allen Brain Atlas.}
\format{
An object of class \code{"data.frame"}.
}
\source{
Allen Brain Atlas
}
\usage{
data(abagenes)
}
\description{
gene identifiers from ABA
}
\details{
\itemize{
\item hgnc_genes. The SYMBOL of each gene.
\item entrezgene. entrezgene id.
\item ensembl_gene_id. ensembl identifier.
}
}
\examples{
data(abagenes)
abagenes
}
\keyword{datasets}
