% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardinas.R
\name{pardinas}
\alias{pardinas}
\title{Output the list of schizophrenia associated genes from Pardinas (PMID:29483656)}
\usage{
pardinas()
}
\value{
a tibble of schizophreina genes with one column labelled 'genes'
}
\description{
Output the list of schizophrenia associated genes from Pardinas (PMID:29483656)
}
\examples{
pardinas()
}
