% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardinas_snp.R
\name{pardinas_snp}
\alias{pardinas_snp}
\title{Output the list of schizophrenia associated snps from Pardinas (PMID:29483656)}
\usage{
pardinas_snp()
}
\value{
a tibble of schizophreina snps with a column labelled 'snps' and a column names 'genes
}
\description{
Output the list of schizophrenia associated snps from Pardinas (PMID:29483656)
}
\examples{
pardinas_snp()
}
