% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripke_snp.R
\name{ripke_snp}
\alias{ripke_snp}
\title{Output the list of schizophrenia associated snps from Ripke (PMID:13595)}
\usage{
ripke_snp()
}
\value{
a tibble of schizophrenia snps with a column labelled 'snps' and a column names 'genes
}
\description{
Output the list of schizophrenia associated snps from Ripke (PMID:13595)
}
\examples{
ripke_snp()
}
